WIP: Change worker id type to a random uuid form
