+++ .env.example
@@ -9,6 +9,6 @@ DRIVER=docker
 WORKER_ADDRESS=tcp://localhost:5555
 
 KEYS_PATH=/home/user/.ssh
-WHITE_LIST_PATH=/home/user/go/src/github.com/ufcg-lsd/arrebol-pb/whitelist
+ALLOWED_LIST_PATH=/home/user/go/src/github.com/ufcg-lsd/arrebol-pb/whitelist
 ARREBOL_PRIV_KEY_PATH=/home/user/arrebol_key
 ARREBOL_PUB_KEY_PATH=/home/user/arrebol_key.pem
