+++ storage/workers_test.go
@@ -1,0 +1,7 @@
+package storage
+
+import "testing"
+
+func TestUUID(t *testing.T) {
+
+}
